Add posibility to specify main script
If there is an existing main.sh execute it in parallel to supervisord
and exit the container as soon as this script exits.
Supervirsod seems not to have a method to exit on it's own when only
one(!) services exits, methods i found (e.g. [1]) seem to only be able
to treat all processes the same.

Idea taken from [2].

[1] https://gist.github.com/tomazzaman/63265dfab3a9a61781993212fa1057cb
[2] https://github.com/oott123/docker-novnc